+++ packages/bs4/src/binders/scroll-to-on-event.binder.ts
@@ -1,36 +1,4 @@
-import { IBinder } from '@ribajs/core';
-
-/**
- * Scrolls to an element by event and selector
- *
- * Attributes:
- *  * scroll-element="query-selector"
- * @see https://stackoverflow.com/a/31987330
- * @param element
- * @param to
- * @param duration
- */
-const scrollTo = (to: HTMLElement, offset: number, scrollElement: Element | (Window & typeof globalThis) | null) => {
-  if (!scrollElement) {
-    scrollElement = window;
-  }
-
-  if (typeof((scrollElement as Window).pageYOffset) === 'number') {
-    // if is is window to scroll
-    scrollElement.scroll({
-      behavior: 'smooth',
-      left: 0,
-      top: (to.getBoundingClientRect().top + (scrollElement as Window).pageYOffset) - offset,
-    });
-  } else {
-    // if is is another element to scroll
-    scrollElement.scroll({
-      behavior: 'smooth',
-      left: 0,
-      top: (to.offsetTop ) - offset,
-    });
-  }
-};
+import { IBinder, Utils } from '@ribajs/core';
 
 export const scrollToOnEventBinder: IBinder<string> = {
   name: 'scroll-to-on-*',
@@ -39,7 +7,7 @@ export const scrollToOnEventBinder: IBinder<string> = {
     this.customData.onEvent = (event: Event) => {
       const offset = Number(el.dataset.offset || 0);
       const scrollElement = el.dataset.scrollElement ? document.querySelector(el.dataset.scrollElement) : window;
-      scrollTo(this.customData.target, offset, scrollElement);
+      Utils.scrollTo(this.customData.target, offset, scrollElement);
       event.preventDefault();
     };
     const eventName = this.args[0] as string;