+++ packages/router/src/services/View.ts
@@ -3,30 +3,20 @@ import { IState } from '../interfaces';
 
 /**
  * BaseView to be extended
- *
- * @namespace Barba.BaseView
- * @type {Object}
  */
 abstract class BaseView {
   /**
    * Namespace of the view.
    * (need to be associated with the data-namespace of the container)
-   *
-   * @memberOf Barba.BaseView
-   * @type {string}
    */
   protected namespace?: string;
 
-  protected $container?: JQuery<HTMLElement>;
+  protected container?: HTMLElement;
 
   private dispatcher = new EventDispatcher();
 
   /**
    * Helper to extend the object
-   *
-   * @memberOf Barba.BaseView
-   * @param  {Object} newObject
-   * @return {Object} newInheritObject
    */
   public extend(obj: object) {
     return Utils.extend(false, this, obj);
@@ -37,8 +27,6 @@ abstract class BaseView {
    * P.S. Is suggested to init the view before starting Barba.Pjax.start(),
    * in this way .onEnter() and .onEnterCompleted() will be fired for the current
    * container when the page is loaded.
-   *
-   * @memberOf Barba.BaseView
    */
   public init() {
     const self = this;
@@ -49,8 +37,8 @@ abstract class BaseView {
       }
     });
 
-    this.dispatcher.on('newPageReady', (viewId: string, newStatus: IState, oldStatus: IState, $container: JQuery<HTMLElement>, html: string, isInit: boolean) => {
-      self.$container = $container;
+    this.dispatcher.on('newPageReady', (viewId: string, newStatus: IState, oldStatus: IState, container: HTMLElement, html: string, isInit: boolean) => {
+      self.container = container;
       if (newStatus.namespace === self.namespace) {
         self.onEnter();
       }
@@ -70,36 +58,24 @@ abstract class BaseView {
  /**
   * This function will be fired when the container
   * is ready and attached to the DOM.
-  *
-  * @memberOf Barba.BaseView
-  * @abstract
   */
  protected abstract onEnter(): any;
 
   /**
    * This function will be fired when the transition
    * to this container has just finished.
-   *
-   * @memberOf Barba.BaseView
-   * @abstract
    */
   protected abstract onEnterCompleted(): any;
 
   /**
    * This function will be fired when the transition
    * to a new container has just started.
-   *
-   * @memberOf Barba.BaseView
-   * @abstract
    */
   protected abstract onLeave(): any;
 
   /**
    * This function will be fired when the container
    * has just been removed from the DOM.
-   *
-   * @memberOf Barba.BaseView
-   * @abstract
    */
   protected abstract onLeaveCompleted(): any;
 }